begin hw_22.1
