+++ blog/templates/blog/blog_detail.html
@@ -6,7 +6,7 @@
             <div class="col-lg-6 col-md-6 col-sm-12">
                 <div class="card mb-4 box-shadow">
                     <div class="card-body">
-                        <img src="{{object.img|mymedia_filter}}" width="500">
+                        <img src="/media/{{object.img}}" width="500">
                     </div>
                 </div>
             </div>
@@ -21,8 +21,8 @@
                 </div>
             </div>
             <div>
-                <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:update_blog' object.pk %}">Редактировать</a>
-                <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:delete_blog' object.pk %}">Удалить</a>
+                <a class="p-2 btn btn-outline-primary" href="{% url 'blog:update_blog' object.pk %}">Редактировать</a>
+                <a class="p-2 btn btn-outline-primary" href="{% url 'blog:delete_blog' object.pk %}">Удалить</a>
             </div>
         </div>
     </div>
