+++ blog/templates/blog/blog_list.html
@@ -2,7 +2,7 @@
 {% load my_tags %}
 {% block content %}
     <div class="col-12 mb-5">
-        <a class="btn btn-outline-primary" href="{% url 'catalog:create_blog' %}">Создать</a>
+        <a class="btn btn-outline-primary" href="{% url 'blog:create_blog' %}">Создать</a>
     </div>
     <div class="row text-center">
         {% for object in object_list %}
@@ -14,9 +14,9 @@
                     <div class="card-body">
                         <img src="{% mymedia_tag object.img %}" width="100">
                         <div>
-                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:blog' object.pk %}">Просмотр</a>
-                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:update_blog' object.pk %}">Редактировать</a>
-                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:delete_blog' object.pk %}">Удалить</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'blog:blog' object.pk %}">Просмотр</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'blog:update_blog' object.pk %}">Редактировать</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'blog:delete_blog' object.pk %}">Удалить</a>
                         </div>
                     </div>
                 </div>
