+++ catalog/forms.py
@@ -1,0 +1,29 @@
+from django import forms
+
+from catalog.models import Product
+
+
+class ProductForm(forms.ModelForm):
+
+    class Meta:
+        model = Product
+        fields = ('name', 'description', 'img', 'category', 'price')
+
+    def clean_name(self):
+        cleaned_data = self.cleaned_data.get('name')
+        ban_words = ['казино', 'криптовалюта', 'крипта', 'биржа', 'дешево', 'бесплатно', 'обман', 'полиция', 'радар']
+
+        for word in ban_words:
+            if word in cleaned_data:
+                raise forms.ValidationError('Используются запретные слова')
+        return cleaned_data
+
+    def clean_description(self):
+        cleaned_data = self.cleaned_data.get('description')
+        ban_words = ['казино', 'криптовалюта', 'крипта', 'биржа', 'дешево', 'бесплатно', 'обман', 'полиция', 'радар']
+
+        for word in ban_words:
+            if word in cleaned_data:
+                raise forms.ValidationError('Используются запретные слова')
+        return cleaned_data
+
