+++ catalog/models.py
@@ -18,7 +18,7 @@ class Category(models.Model):
 
 
 class Product(models.Model):
-    name = models.CharField(max_length=100, verbose_name='Наименование')
+    name = models.CharField(max_length=100, verbose_name='Наименование', unique=True)
     description = models.TextField(**NULLABLE, verbose_name='Описание')
     img = models.ImageField(upload_to='product/', **NULLABLE, verbose_name='Изображение')
     category = models.ForeignKey(Category, on_delete=models.CASCADE, verbose_name='Категория')
@@ -32,20 +32,3 @@ class Product(models.Model):
     class Meta:
         verbose_name = 'продукт'
         verbose_name_plural = 'продукты'
-
-
-class Blog(models.Model):
-    title = models.CharField(max_length=150, verbose_name='Заголовок')
-    slug = models.CharField(max_length=150, verbose_name='Slug')
-    text = models.TextField(verbose_name='Содержимое', **NULLABLE)
-    img = models.ImageField(upload_to='blog/', **NULLABLE, verbose_name='Изображение')
-    date_begin = models.DateField(default=date.today, verbose_name='Дата создания')
-    activate = models.BooleanField(default=True)
-    count_views = models.IntegerField(default=0, verbose_name='Количество просмотров')
-
-    def __str__(self):
-        return f'{self.title}'
-
-    class Meta:
-        verbose_name = 'блог'
-        verbose_name_plural = 'блоги'
