+++ catalog/templates/catalog/category_list.html
@@ -1,6 +1,9 @@
 {% extends 'catalog/base.html' %}
 {% load my_tags %}
 {% block content %}
+    <div class="col-12 mb-5">
+        <a class="btn btn-primary" href="{% url 'catalog:create_product' %}">Добавить товар</a>
+    </div>
     <div class="row text-center">
         {% for object in object_list %}
             <div class="col-3">
@@ -11,7 +14,7 @@
                     <div class="card-body">
                         <img src="{% mymedia_tag object.img %}" width="100">
                         <div>
-                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:products' object.pk %}">К товарам</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:product_list' object.pk %}">К товарам</a>
                         </div>
                     </div>
                 </div>
