+++ catalog/templates/catalog/product_confirm_delete.html
@@ -1,0 +1,21 @@
+{% extends 'catalog/base.html' %}
+
+{% block content %}
+    <div class="col-12">
+        <div class="row">
+            <div class="col-6">
+                <div class="card">
+                    <div class="card-body">
+                        <form method="post">
+                            {% csrf_token %}
+                            <p> Хотите удалить товар"{{ object.title }}"?</p>
+                            <button type="submit" class="btn btn-primary">Подтвердить</button>
+                            <a href="{% url 'catalog:category' %}" class="btn btn-primary">Отмена</a>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+{% endblock %}
