+++ catalog/templates/catalog/product_detail.html
@@ -6,7 +6,7 @@
             <div class="col-lg-6 col-md-6 col-sm-12">
                 <div class="card mb-4 box-shadow">
                     <div class="card-body">
-                        <img src="{{object.img|mymedia_filter}}" >
+                        <img src="{{object.img|mymedia_filter}}" width="500">
                     </div>
                 </div>
             </div>
