+++ catalog/templates/catalog/product_form.html
@@ -1,0 +1,21 @@
+{% extends 'catalog/base.html' %}
+
+{% block content %}
+    <div class="col-6">
+        <div class="card-header">
+            <h3 class="card-title">
+                Добавление товара
+            </h3>
+        </div>
+        <div class="card-body">
+            <form method="post" enctype="multipart/form-data">
+                {% csrf_token %}
+                {{ form.as_p }}
+                <button type="submit" class="btn btn-success">Сохранить</button>
+            </form>
+        </div>
+        <div class="card-footer"></div>
+    </div>
+
+
+{% endblock %}
