+++ catalog/templates/catalog/product_list.html
@@ -10,7 +10,11 @@
                     </div>
                     <div class="card-body">
                         <img src="{% mymedia_tag object.img %}" width="100">
-                        <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:product' object.pk %}">Описание</a>
+                        <div>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:product' object.pk %}">Описание</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:update_product' object.pk %}">Редактировать</a>
+                            <a class="p-2 btn btn-outline-primary" href="{% url 'catalog:delete_product' object.pk %}">Удалить</a>
+                        </div>
                     </div>
                 </div>
             </div>
